refactor: share tty
