+++ config/config.go
@@ -52,6 +52,7 @@ type ServerConfig struct {
 	Term            string `default:"xterm"`
 	ShowLocation    bool
 	EnableShare     bool
+	Collabrate      bool
 
 	// audit
 	EnableAudit bool
