+++ route/handler.go
@@ -2,169 +2,19 @@ package route
 
 import (
 	"bytes"
-	"context"
 	"encoding/json"
 	"fmt"
 	"net/http"
 	"net/url"
-	"strings"
-	"time"
 
 	"github.com/gin-gonic/gin"
 	"github.com/gorilla/websocket"
 	log "github.com/sirupsen/logrus"
 	"github.com/yudai/gotty/webtty"
 
-	"github.com/wrfly/container-web-tty/audit"
 	"github.com/wrfly/container-web-tty/types"
+	"github.com/wrfly/container-web-tty/util"
 )
-
-func (server *Server) handleExec(c *gin.Context, counter *counter) {
-	cInfo := server.containerCli.GetInfo(c.Request.Context(), c.Param("id"))
-	server.generateHandleWS(c.Request.Context(), counter, cInfo).
-		ServeHTTP(c.Writer, c.Request)
-}
-
-func (server *Server) generateHandleWS(ctx context.Context, counter *counter, container types.Container) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		if container.Shell == "" {
-			log.Errorf("cannot find a valid shell in container [%s]", container.ID)
-			return
-		}
-
-		num := counter.add(1)
-		closeReason := "unknown reason"
-
-		defer func() {
-			num := counter.done()
-			if strings.Contains(closeReason, "error") {
-				log.Errorf("Connection closed by %s: %s, connections: %d",
-					closeReason, r.RemoteAddr, num)
-			}
-			log.Infof("Connection closed by %s: %s, connections: %d",
-				closeReason, r.RemoteAddr, num)
-		}()
-
-		if int64(server.options.MaxConnection) != 0 {
-			if num > server.options.MaxConnection {
-				closeReason = "exceeding max number of connections"
-				return
-			}
-		}
-
-		log.Infof("New client connected: %s, connections: %d", r.RemoteAddr, num)
-
-		conn, err := server.upgrader.Upgrade(w, r, nil)
-		if err != nil {
-			closeReason = err.Error()
-			return
-		}
-		defer conn.Close()
-
-		cctx, timeoutCancel := context.WithCancel(ctx)
-		defer timeoutCancel()
-
-		err = server.processTTY(cctx, timeoutCancel, conn, container)
-		switch err {
-		case ctx.Err():
-			closeReason = "cancelation"
-		case cctx.Err():
-			closeReason = "time out"
-		case webtty.ErrSlaveClosed:
-			closeReason = "backend closed"
-		case webtty.ErrMasterClosed:
-			closeReason = "tab closed"
-		default:
-			closeReason = fmt.Sprintf("an error: %s", err)
-		}
-	}
-}
-
-func (server *Server) processTTY(ctx context.Context, timeoutCancel context.CancelFunc,
-	conn *websocket.Conn, container types.Container) error {
-	arguments, err := server.readInitMessage(conn)
-	if err != nil {
-		return err
-	}
-	log.Debugf("exec container: %s, params: %s", container.ID, arguments)
-
-	q, err := parseQuery(strings.TrimSpace(arguments))
-	if err != nil {
-		return err
-	}
-	container.Exec = types.ExecOptions{
-		Cmd:        q.Get("cmd"),
-		Env:        q.Get("env"),
-		User:       q.Get("user"),
-		Privileged: q.Get("p") != "",
-	}
-
-	containerTTY, err := server.containerCli.Exec(ctx, container)
-	if err != nil {
-		return fmt.Errorf("exec container error: %s", err)
-	}
-	defer containerTTY.Exit()
-
-	// handle timeout
-	tout := server.options.IdleTime
-	if tout.Seconds() != 0 {
-		go func() {
-			timer := time.NewTimer(tout)
-			activeChan := containerTTY.ActiveChan()
-			for {
-				select {
-				case <-timer.C:
-					timer.Stop()
-					timeoutCancel()
-					return
-				case <-activeChan:
-					// the connection is active, reset the timer
-					timer.Reset(tout)
-				}
-			}
-
-		}()
-	}
-
-	titleBuf, err := server.makeTitleBuff(container)
-	if err != nil {
-		return fmt.Errorf("failed to fill window title template: %s", err)
-	}
-
-	opts := []webtty.Option{
-		webtty.WithWindowTitle(titleBuf),
-		webtty.WithPermitWrite(),
-		// webtty.WithReconnect(10), // not work....
-	}
-
-	wrapper := &wsWrapper{conn}
-	shareableTTY := types.NewShareTTY(containerTTY)
-	server.mMux.Lock()
-	server.masters[container.ID] = shareableTTY
-	server.mMux.Unlock()
-	defer func() {
-		server.mMux.Lock()
-		delete(server.masters, container.ID)
-		server.mMux.Unlock()
-	}()
-
-	if server.options.EnableAudit {
-		cIP := conn.RemoteAddr().String()
-		r := shareableTTY.Fork(cIP)
-		go audit.LogTo(ctx, r, audit.LogOpts{
-			Dir:         server.options.AuditLogDir,
-			ContainerID: container.ID,
-			ClientIP:    cIP,
-		})
-	}
-
-	tty, err := webtty.New(wrapper, shareableTTY, opts...)
-	if err != nil {
-		return fmt.Errorf("failed to create webtty: %s", err)
-	}
-
-	return tty.Run(ctx)
-}
 
 func (server *Server) handleWSIndex(c *gin.Context) {
 	cInfo := server.containerCli.GetInfo(c.Request.Context(), c.Param("id"))
@@ -340,7 +190,7 @@ func (server *Server) handleLogs(c *gin.Context) {
 
 	tty, err := webtty.New(
 		&wsWrapper{conn},
-		newSlave(logsReadCloser, false),
+		newSlave(util.NopRWCloser(logsReadCloser), false),
 		[]webtty.Option{
 			webtty.WithWindowTitle(titleBuf),
 			webtty.WithPermitWrite(), // can type "enter"
@@ -355,62 +205,6 @@ func (server *Server) handleLogs(c *gin.Context) {
 		if err != webtty.ErrMasterClosed && err != webtty.ErrSlaveClosed {
 			log.Errorf("failed to run webtty: %s", err)
 		}
-	}
-}
-
-func (server *Server) handleShare(c *gin.Context) {
-	ctx := c.Request.Context()
-	cid := c.Param("id")
-	cInfo := server.containerCli.GetInfo(ctx, cid)
-
-	conn, err := server.upgrader.Upgrade(c.Writer, c.Request, nil)
-	if err != nil {
-		log.Errorf("upgrade ws error: %s", err)
-		return
-	}
-	defer conn.Close()
-
-	// note: must read the init message
-	// although it's useless in this situation
-	server.readInitMessage(conn)
-
-	server.mMux.RLock()
-	shareableTTY, ok := server.masters[cInfo.ID]
-	server.mMux.RUnlock()
-	if !ok {
-		log.Error("share terminal error, master not found")
-		conn.WriteMessage(websocket.CloseMessage, []byte("not found"))
-		return
-	}
-
-	titleBuf, err := server.makeTitleBuff(cInfo)
-	if err != nil {
-		e := fmt.Sprintf("failed to fill window title template: %s", err)
-		conn.WriteMessage(websocket.CloseMessage, []byte(e))
-		log.Error(e)
-		return
-	}
-
-	fork := shareableTTY.Fork(c.ClientIP())
-	defer fork.Close()
-
-	tty, err := webtty.New(
-		&wsWrapper{conn},
-		newSlave(fork, true),
-		[]webtty.Option{
-			webtty.WithWindowTitle(titleBuf),
-			webtty.WithPermitWrite()}...,
-	)
-	if err != nil {
-		e := fmt.Sprintf("failed to create webtty: %s", err)
-		conn.WriteMessage(websocket.CloseMessage, []byte(e))
-		log.Error(e)
-		return
-	}
-
-	if err := tty.Run(ctx); err != nil && err != webtty.ErrMasterClosed {
-		e := fmt.Sprintf("failed to run webtty: %s", err)
-		log.Error(e)
 	}
 }
