+++ route/slave_wrapper.go
@@ -1,16 +1,13 @@
 package route
 
 import (
-	"bytes"
 	"io"
 
 	"github.com/yudai/gotty/webtty"
 )
 
 type slaveWrapper struct {
-	rc io.ReadCloser
-	pr io.ReadCloser
-	pw io.WriteCloser
+	master io.ReadWriteCloser
 }
 
 func (sw *slaveWrapper) WindowTitleVariables() map[string]interface{} {
@@ -22,62 +19,61 @@ func (sw *slaveWrapper) ResizeTerminal(columns int, rows int) error {
 }
 
 func (sw *slaveWrapper) Write(p []byte) (n int, err error) {
-	if p[0] == 13 {
-		p = append(p, 10) // append a new line
-	}
-	return sw.pw.Write(p)
+	// logrus.Debugf("slaveWrapper write-> %s", p)
+	// if p[0] == 13 {
+	// 	p = append(p, 10) // append a new line
+	// }
+	return sw.master.Write(p)
 }
 
 func (sw *slaveWrapper) Read(p []byte) (n int, err error) {
-	return sw.pr.Read(p)
+	return sw.master.Read(p)
 }
 
-func newSlave(rc io.ReadCloser, share bool) webtty.Slave {
-	pr, pw := io.Pipe()
-	go func() {
-		defer pr.Close()
-		defer pw.Close()
-		defer rc.Close()
+func newSlave(master io.ReadWriteCloser, share bool) webtty.Slave {
+	// pr, pw := io.Pipe()
+	// go func() {
+	// 	defer pr.Close()
+	// 	defer pw.Close()
+	// 	defer master.Close()
 
-		bs := make([]byte, 2048)
-		for {
-			n, err := rc.Read(bs)
-			if err != nil {
-				return
-			}
-			if share {
-				pw.Write(bs[:n])
-				continue
-			}
+	// 	bs := make([]byte, 2048)
+	// 	for {
+	// 		n, err := master.Read(bs)
+	// 		if err != nil {
+	// 			return
+	// 		}
+	// 		if share {
+	// 			pw.Write(bs[:n])
+	// 			continue
+	// 		}
+	// 		panic(1)
+	// 		if n <= 1 {
+	// 			continue
+	// 		}
 
-			if n <= 1 {
-				continue
-			}
+	// 		if n >= 2 && bs[n-2] == 13 { // \r\n
+	// 			pw.Write(bs[:n])
+	// 			continue
+	// 		}
 
-			if n >= 2 && bs[n-2] == 13 { // \r\n
-				pw.Write(bs[:n])
-				continue
-			}
-
-			// only \n or a long log string contains \n
-			s, e := 0, 0
-			for e < n {
-				x := bytes.IndexByte(bs[e:n], 10)
-				if x == -1 {
-					break
-				}
-				s = e
-				e += x
-				pw.Write(bs[s:e])
-				pw.Write([]byte{13, 10})
-				e++ // skip this \n
-			}
-		}
-	}()
+	// 		// only \n or a long log string contains \n
+	// 		s, e := 0, 0
+	// 		for e < n {
+	// 			x := bytes.IndexByte(bs[e:n], 10)
+	// 			if x == -1 {
+	// 				break
+	// 			}
+	// 			s = e
+	// 			e += x
+	// 			pw.Write(bs[s:e])
+	// 			pw.Write([]byte{13, 10})
+	// 			e++ // skip this \n
+	// 		}
+	// 	}
+	// }()
 
 	return &slaveWrapper{
-		rc: rc,
-		pr: pr,
-		pw: pw,
+		master: master,
 	}
 }
